+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ENHANCEMENTS:
 * add resource junos_security_utm_policy
 * add resource junos_security_utm_profile_web_filtering_juniper_enhanced
 * add resource junos_security_utm_profile_web_filtering_juniper_local
+* add resource junos_security_utm_profile_web_filtering_websense_redirect
 
 BUG FIXES:
 * [workflows] fix compile freebsd/arm64 on release